+++ r/src/matrix_utils.cpp
@@ -10,6 +10,7 @@
 #define RCPP_NO_SUGAR
 #include <Rcpp.h>
 #include <RcppEigen.h>
+#include "bpcells-cpp/matrixIterators/BernoulliSample.h"
 #include "bpcells-cpp/matrixIterators/CSparseMatrix.h"
 #include "bpcells-cpp/matrixIterators/ColwiseRank.h"
 #include "bpcells-cpp/matrixIterators/MatrixAddition.h"
@@ -343,6 +344,33 @@ SEXP iterate_matrix_mask_double_cpp(SEXP mat, SEXP mask, bool invert) {
             take_unique_xptr<MatrixLoader<double>>(mat),
             take_unique_xptr<MatrixLoader<uint32_t>>(mask)
         );
+}
+
+// [[Rcpp::export]]
+SEXP iterate_matrix_bernoulli_sample_uint32_t_cpp(SEXP mat, double prob, int seed) {
+    return make_unique_xptr<BernoulliSample<uint32_t>>(
+        take_unique_xptr<MatrixLoader<uint32_t>>(mat),
+        prob,
+        static_cast<uint64_t>(static_cast<uint32_t>(seed))
+    );
+}
+
+// [[Rcpp::export]]
+SEXP iterate_matrix_bernoulli_sample_float_cpp(SEXP mat, double prob, int seed) {
+    return make_unique_xptr<BernoulliSample<float>>(
+        take_unique_xptr<MatrixLoader<float>>(mat),
+        prob,
+        static_cast<uint64_t>(static_cast<uint32_t>(seed))
+    );
+}
+
+// [[Rcpp::export]]
+SEXP iterate_matrix_bernoulli_sample_double_cpp(SEXP mat, double prob, int seed) {
+    return make_unique_xptr<BernoulliSample<double>>(
+        take_unique_xptr<MatrixLoader<double>>(mat),
+        prob,
+        static_cast<uint64_t>(static_cast<uint32_t>(seed))
+    );
 }
 
 // [[Rcpp::export]]